admin/requirements.txt
asgiref==3.7.2
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
pika==1.3.2